Improve tests classification in e2e, integr, unit
Last rearrangement went a long way, but there were still some unit tests
in e2e. This arrangement ensures that tests in
tests/e2e_integration/cli_commands are e2e tests that start and end with
CLI commands.

Regarding the latter: this commits also reworks the helper function for
waiting for task completion to be a CLI-based helper so that it can be
used for CLI polling. The function now uses
divbase-cli task-history id to poll for task terminal status
(SUCCESS or FAILURE)
